Fixed error when the currently logged on user is in a list
When a user tried to a follow a list of people and the user is in that list, it would error out. There is an error checking process on the array of twitter IDs. The client ID is gathered and stored as a variable and then a delete_if is ran against the array of new people. If the IDs are a match, then it's removed from the array.

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>